requirements.txt
mistralai 
cartesia
python-dotenv
fastapi 
uvicorn 
python-multipart